Added Class to DbContext, fixed UCD-Plan, fixed itemtype and typeitem doubled class (modifying UCD-PurchaseItem) and fixed attribute types for Bizum, CreditCard and Paypal. (#18)

+++ src/AppForSEII2526.API/Models/Bizum.cs
@@ -2,6 +2,6 @@
 {
     public class Bizum:PaymentMethod
     {
-        public int TelephoneNumber { get; set; }
+        public long TelephoneNumber { get; set; }
     }
 }
